% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_df.R
\name{print_df}
\alias{print_df}
\title{Automatically Format printing for dataframes}
\usage{
print_df(x, ...)
}
\arguments{
\item{x}{a dataframe}

\item{...}{an ellispse}
}
\value{
a dataframe
}
\description{
Automatically Format printing for dataframes
}
\author{
Daniel Hintz
}
