% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_df_hlp.R
\name{print_df_hlp}
\alias{print_df_hlp}
\title{Helper functon for print_df}
\usage{
print_df_hlp(x, ...)
}
\arguments{
\item{x}{a dataframe}

\item{...}{an ellipse}
}
\value{
a formated dataframe
}
\description{
Helper functon for print_df
}
\author{
Daniel Hintz
}
