% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_vec.R
\name{print_vec}
\alias{print_vec}
\title{Automated Print Formatting for matrices}
\usage{
print_vec(x, ...)
}
\arguments{
\item{x}{A matrix}

\item{...}{ellispe not set}
}
\value{
A matrix
}
\description{
Automated Print Formatting for matrices
}
\author{
Daniel Hintz
}
