% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_vec_hlp.R
\name{print_vec_hlp}
\alias{print_vec_hlp}
\title{Helper function for print_vec}
\usage{
print_vec_hlp(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{an ellipse}
}
\description{
Helper function for print_vec
}
\author{
Daniel Hintz
}
