% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{tex}
\alias{tex}
\title{Copy and paste solution for putting common Latex code in script}
\usage{
tex()
}
\value{
a character vector
}
\description{
Copy and paste solution for putting common Latex code in script
}
\author{
Daniel Hintz
}
