% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_docs.R
\name{tex_docs}
\alias{tex_docs}
\title{Download template preamble.tex and .bib files for pdf Rmarkdown}
\usage{
tex_docs(pream = TRUE, bib = TRUE)
}
\arguments{
\item{pream}{boolean. download preamble.tex}

\item{bib}{boolean. download references.bib}
}
\description{
Download template preamble.tex and .bib files for pdf Rmarkdown
}
