% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_LV_inc.R
\name{spell_LV_inc}
\alias{spell_LV_inc}
\title{spell_LV_inc}
\usage{
spell_LV_inc(dat, corr, t_low = 3, t_high = 4, ...)
}
\arguments{
\item{dat}{a character vector with misspellings to be corrected}

\item{corr}{a character vector with correct spellings for reference}

\item{t_low}{a number to set level for the more sensitive thresholds}

\item{t_high}{a number to set level for the less sensitive thresholds}

\item{...}{ellipse for end user flexibility}
}
\value{
a character vector
}
\description{
The \code{spell_LV_inc()} function is used to inspect "INCONCLUSIVE" strings for spell checker performance with two thresholds.
}
\details{
The \code{spell_LV_inc()} function uses Levenshtein distance to calculate the string distance with two different thresholds by which if the threshold is beyond that threshold the string is replaced with "INCONCLUSIVE". This function helps inspect what is deemed "INCONCLUSIVE" and if it can be manually corrected or if a lower threshold classifies all strings correctly.
}
\examples{
dat <- c(NA, "Rawln", NA, NA, "", "Torrinton", "Laraie", "Laamie","Torringtn","Whatld","Rwlins", "(D)", "eryfeb")
corr <- c("Wheatland", "Torrington", "Rawlins", "Laramie")
spell_LV_inc(dat, corr,3, 9)
spell_LV_inc(dat, corr,3, 4)
spell_LV_inc(dat, corr)
spell_LV_inc(dat, corr,2, 4)
dat <- spell_LV(dat, corr,2)
dat[c(2,10)]
dat[c(2,10)] <- c("Rawlins","Wheatland")
}
\seealso{
See documentation for \code{?spell_LV} for underlying mechanics
}
\author{
Daniel Hintz
}
